added issue_template.md to get useful bug reports
This Issue Template will be used automagically by github web when someone creates a new issue. They will see the template text and hopefully fill in the different sections. This helps reducing the "X is broken. Please fix it" issues, where you then have to ask 10 times to get some details, how to reproduce and so on.